adf_util: util. functions for AdfList moved to adf_str.
