+++ src/adf_util.h
@@ -40,11 +40,6 @@ void swLong ( uint8_t * const buf,
 void swShort ( uint8_t * const buf,
                const uint16_t  val );
 
-PREFIX struct AdfList * newCell ( struct AdfList * const list,
-                                  void * const           content );
-
-PREFIX void freeList ( struct AdfList * const list );
-
 PREFIX void adfDays2Date( int32_t       days,
                           int * const   yy,
                           int * const   mm,
